refactor(lib): split types into domain-specific modules
Reorganize monolithic types/index.ts into focused domain files:
- cognitive-level.ts: CognitiveLevel enum and helpers
- components.ts: Component types (Instruction, Knowledge)
- core.ts: Module and base types
- persona.ts: Persona types
- primitives.ts: Primitive types (ProcessStep, Constraint, etc.)
- guards.ts: Type guard functions
- validation.ts: Validation result types
- build-report.ts: Build report types
- registry.ts: Registry types

index.ts now serves as barrel export for backward compatibility.
