+++ packages/ums-lib/src/types/build-report.ts
@@ -1,0 +1,86 @@
+/**
+ * @file Build report types for UMS v2.1.
+ * @description Defines structures for build process reporting and metadata.
+ */
+
+// #region Composition Event
+
+/**
+ * A composition event representing a module replacement or merge operation.
+ */
+export interface CompositionEvent {
+  /** The ID of the module. */
+  id: string;
+  /** The version of the module. */
+  version: string;
+  /** The source of the module. */
+  source: string;
+  /** The SHA-256 digest of the module content. */
+  digest: string;
+  /** The composition strategy used (base or replace). */
+  strategy: 'base' | 'replace';
+}
+
+// #endregion
+
+// #region Build Report Module
+
+/**
+ * A report for a single module within the build.
+ */
+export interface BuildReportModule {
+  /** The ID of the module. */
+  id: string;
+  /** The name of the module. */
+  name: string;
+  /** The version of the module. */
+  version: string;
+  /** A string representation of the module's source. */
+  source: string;
+  /** A SHA-256 digest of the module file content. */
+  digest: string;
+  /** Flag indicating if the module is deprecated. */
+  deprecated: boolean;
+  /** The ID of a successor module, if this module is deprecated. */
+  replacedBy?: string;
+  /** Optional composition history if this module was replaced or merged. */
+  composedFrom?: CompositionEvent[];
+}
+
+// #endregion
+
+// #region Build Report Group
+
+/**
+ * A report for a single module group within the build.
+ */
+export interface BuildReportGroup {
+  /** The name of the module group. */
+  groupName: string;
+  /** A list of modules in this group. */
+  modules: BuildReportModule[];
+}
+
+// #endregion
+
+// #region Build Report
+
+/**
+ * A report generated by the build process, containing metadata about the build.
+ */
+export interface BuildReport {
+  /** The name of the persona that was built. */
+  personaName: string;
+  /** The UMS schema version used for the build. */
+  schemaVersion: string;
+  /** The version of the tool that generated the build. */
+  toolVersion: string;
+  /** A SHA-256 digest of the persona file content. */
+  personaDigest: string;
+  /** The timestamp of the build in ISO 8601 UTC format. */
+  buildTimestamp: string;
+  /** The list of module groups included in the build. */
+  moduleGroups: BuildReportGroup[];
+}
+
+// #endregion
